+++ fmi-import/src/main/kotlin/no/ntnu/ihb/fmi4j/Model.kt
@@ -29,12 +29,6 @@ import java.io.Closeable
 
 interface Model : Closeable {
 
-    val guid: String
-        get() = modelDescription.guid
-
-    val modelName: String
-        get() = modelDescription.modelName
-
     val modelDescription: ModelDescription
 
     fun newInstance(): SlaveInstance
