+++ fmi-import/src/main/kotlin/no/ntnu/ihb/fmi4j/importer/fmi1/Fmu.kt
@@ -24,19 +24,24 @@
 
 package no.ntnu.ihb.fmi4j.importer.fmi1
 
+import no.ntnu.ihb.fmi4j.Model
+import no.ntnu.ihb.fmi4j.SlaveInstance
+import no.ntnu.ihb.fmi4j.importer.AbstractFmu
+import no.ntnu.ihb.fmi4j.importer.fmi1.jni.CoSimulationLibraryWrapper
+import no.ntnu.ihb.fmi4j.importer.fmi1.jni.Fmi1CoSimulationLibrary
 import no.ntnu.ihb.fmi4j.importer.fmi1.jni.Fmi1Library
+import no.ntnu.ihb.fmi4j.modeldescription.CoSimulationModelDescription
 import no.ntnu.ihb.fmi4j.modeldescription.CommonModelDescription
-import no.ntnu.ihb.fmi4j.modeldescription.ModelDescriptionProvider
 import no.ntnu.ihb.fmi4j.modeldescription.fmi1.JaxbModelDescriptionParser
 import no.ntnu.ihb.fmi4j.util.OsUtil
 import no.ntnu.ihb.fmi4j.util.extractContentTo
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
-import java.io.*
+import java.io.ByteArrayInputStream
+import java.io.File
+import java.io.FileNotFoundException
+import java.io.IOException
 import java.net.URL
-import java.nio.file.Files
-import java.util.*
-import java.util.concurrent.atomic.AtomicBoolean
 
 /**
  * Represents an FMU
@@ -44,70 +49,36 @@ import java.util.concurrent.atomic.AtomicBoolean
  * @author Lars Ivar Hatledal
  */
 class Fmu private constructor(
-        val name: String,
-        private val extractedFmu: File
-) : Closeable {
+        name: String,
+        extractedFmu: File
+) : AbstractFmu(name, extractedFmu), Model {
 
-    private var isClosed = AtomicBoolean(false)
     private val libraries = mutableListOf<Fmi1Library>()
     private val instances = mutableListOf<AbstractFmuInstance<*, *>>()
 
-    private val coSimulationFmu by lazy {
-        CoSimulationFmu(this)
-    }
-
-    init {
-        if (!File(extractedFmu, MODEL_DESC).exists()) {
-            deleteExtractedFmuFolder().also {
-                throw IllegalStateException("FMU is invalid, no $MODEL_DESC present!")
-            }
-        }
-
-        synchronized(fmus) {
-            fmus.add(this)
+    private val lib: Fmi1CoSimulationLibrary by lazy {
+        val modelIdentifier = modelDescription.attributes.modelIdentifier
+        val libName = getAbsoluteLibraryPath(modelIdentifier)
+        Fmi1CoSimulationLibrary(libName, modelIdentifier).also {
+            registerLibrary(it)
         }
     }
 
-    val guid: String
-        get() = modelDescription.guid
-
-    val modelName: String
-        get() = modelDescription.modelName
-
-    /**
-     * Get the content of the modelDescription.xml file as a String
-     */
-    val modelDescriptionXml: String
-        get() = modelDescriptionFile.readText()
-
-    val modelDescription: ModelDescriptionProvider by lazy {
-        JaxbModelDescriptionParser.parse(modelDescriptionXml)
+    override fun newInstance(): SlaveInstance {
+        return newInstance(visible = false, interactive = false, loggingOn = false)
     }
 
-    /**
-     * Does the FMU support Co-simulation?
-     */
-    val supportsCoSimulation: Boolean
-        get() = modelDescription.supportsCoSimulation
-
-    /**
-     * Does the FMU support Model Exchange?
-     */
-    val supportsModelExchange: Boolean
-        get() = modelDescription.supportsModelExchange
-
-    fun asCoSimulationFmu(): CoSimulationFmu {
-        if (!supportsCoSimulation) {
-            throw IllegalStateException("FMU does not support Co-simulation!")
+    fun newInstance(visible: Boolean = false, interactive: Boolean = false, loggingOn: Boolean = false): CoSimulationSlave {
+        val c = instantiate(modelDescription, lib, visible, interactive, loggingOn)
+        val wrapper = CoSimulationLibraryWrapper(c, lib)
+        return CoSimulationSlave(wrapper, modelDescription).also {
+            registerInstance(it)
         }
-        return coSimulationFmu
     }
 
-    /**
-     * Get the file handle for the modelDescription.xml file
-     */
-    private val modelDescriptionFile: File
-        get() = File(extractedFmu, MODEL_DESC)
+    override val modelDescription: CoSimulationModelDescription by lazy {
+        JaxbModelDescriptionParser.parse(modelDescriptionXml).asCoSimulationModelDescription()
+    }
 
     private val fmuPath: String
         get() = "file:///${extractedFmu.absolutePath.replace("\\", "/")}"
@@ -115,40 +86,27 @@ class Fmu private constructor(
     /**
      * Get the absolute name of the native library on the form "C://folder/name.extension"
      */
-    fun getAbsoluteLibraryPath(modelIdentifier: String): String {
+    private fun getAbsoluteLibraryPath(modelIdentifier: String): String {
         return File(extractedFmu, BINARIES_FOLDER + File.separator + OsUtil.libraryFolderName + OsUtil.platformBitness
                 + File.separator + modelIdentifier + "." + OsUtil.libExtension).absolutePath
     }
 
-    internal fun registerLibrary(library: Fmi1Library) {
+    private fun registerLibrary(library: Fmi1Library) {
         libraries.add(library)
     }
 
-    internal fun registerInstance(instance: AbstractFmuInstance<*, *>) {
+    private fun registerInstance(instance: AbstractFmuInstance<*, *>) {
         instances.add(instance)
     }
 
-    internal fun instantiate(modelDescription: CommonModelDescription, library: Fmi1Library, visible: Boolean, interactive: Boolean, loggingOn: Boolean): Long {
+    private fun instantiate(modelDescription: CommonModelDescription, library: Fmi1Library, visible: Boolean, interactive: Boolean, loggingOn: Boolean): Long {
         LOG.trace("Calling instantiate: visible=$visible, interactive=$interactive, loggingOn=$loggingOn")
 
         return library.instantiate(modelDescription.attributes.modelIdentifier,
                 modelDescription.guid, fmuPath, visible, interactive, loggingOn)
     }
 
-    override fun close() {
-        if (!isClosed.getAndSet(true)) {
-
-            LOG.debug("Closing FMU '$extractedFmu'..")
-
-            terminateInstances()
-            disposeNativeLibraries()
-            deleteExtractedFmuFolder()
-
-            fmus.remove(this)
-        }
-    }
-
-    private fun disposeNativeLibraries() {
+    override fun disposeNativeLibraries() {
         libraries.forEach {
             it.close()
         }
@@ -156,7 +114,7 @@ class Fmu private constructor(
         System.gc()
     }
 
-    private fun terminateInstances() {
+    override fun terminateInstances() {
         instances.forEach { instance ->
             if (!instance.isTerminated) {
                 instance.terminate()
@@ -168,61 +126,9 @@ class Fmu private constructor(
         instances.clear()
     }
 
-    /**
-     * Deletes the temporary folder where the FMU was extracted
-     */
-    private fun deleteExtractedFmuFolder(): Boolean {
-
-        if (extractedFmu.exists()) {
-            return extractedFmu.deleteRecursively().also { success ->
-                if (success) {
-                    LOG.debug("Deleted extracted FMU contents: $extractedFmu")
-                } else {
-                    LOG.debug("Failed to delete extracted FMU contents: $extractedFmu")
-                }
-            }
-        }
-        return true
-    }
-
-    override fun toString(): String {
-        return "Fmu(fmu=${extractedFmu.absolutePath})"
-    }
-
     companion object {
 
         private val LOG: Logger = LoggerFactory.getLogger(Fmu::class.java)
-
-        private const val RESOURCES_FOLDER = "resources"
-        private const val BINARIES_FOLDER = "binaries"
-
-        private const val FMU_EXTENSION = "fmu"
-        private const val FMI4J_FILE_PREFIX = "fmi4j_"
-
-        private const val MODEL_DESC = "modelDescription.xml"
-
-        private val fmus = Collections.synchronizedList(mutableListOf<Fmu>())
-
-        init {
-            Runtime.getRuntime().addShutdownHook(Thread {
-                synchronized(fmus) {
-                    fmus.toMutableList().forEach {
-                        //mutableList because the list is modified during call
-                        it?.close()
-                    }
-                }
-            })
-        }
-
-        private fun createTempDir(fmuName: String): File {
-            return Files.createTempDirectory(FMI4J_FILE_PREFIX + fmuName).toFile().also {
-                File(it, RESOURCES_FOLDER).apply {
-                    if (!exists()) {
-                        mkdir()
-                    }
-                }
-            }
-        }
 
         /**
          * Creates an FMU from the provided File
