+++ fmi-import/src/main/kotlin/no/ntnu/ihb/fmi4j/importer/fmi2/Fmu.kt
@@ -24,6 +24,7 @@
 
 package no.ntnu.ihb.fmi4j.importer.fmi2
 
+import no.ntnu.ihb.fmi4j.importer.AbstractFmu
 import no.ntnu.ihb.fmi4j.importer.fmi2.jni.Fmi2Library
 import no.ntnu.ihb.fmi4j.modeldescription.CommonModelDescription
 import no.ntnu.ihb.fmi4j.modeldescription.ModelDescriptionProvider
@@ -32,11 +33,11 @@ import no.ntnu.ihb.fmi4j.util.OsUtil
 import no.ntnu.ihb.fmi4j.util.extractContentTo
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
-import java.io.*
+import java.io.ByteArrayInputStream
+import java.io.File
+import java.io.FileNotFoundException
+import java.io.IOException
 import java.net.URL
-import java.nio.file.Files
-import java.util.*
-import java.util.concurrent.atomic.AtomicBoolean
 
 /**
  * Represents an FMU
@@ -44,11 +45,10 @@ import java.util.concurrent.atomic.AtomicBoolean
  * @author Lars Ivar Hatledal
  */
 class Fmu private constructor(
-        val name: String,
-        private val extractedFmu: File
-) : Closeable {
+        name: String,
+        extractedFmu: File
+) : AbstractFmu(name, extractedFmu) {
 
-    private var isClosed = AtomicBoolean(false)
     private val libraries = mutableListOf<Fmi2Library>()
     private val instances = mutableListOf<AbstractFmuInstance<*, *>>()
 
@@ -60,30 +60,7 @@ class Fmu private constructor(
         ModelExchangeFmu(this)
     }
 
-    init {
-        if (!File(extractedFmu, MODEL_DESC).exists()) {
-            deleteExtractedFmuFolder().also {
-                throw IllegalStateException("FMU is invalid, no $MODEL_DESC present!")
-            }
-        }
-        synchronized(fmus) {
-            fmus.add(this)
-        }
-    }
-
-    val guid: String
-        get() = modelDescription.guid
-
-    val modelName: String
-        get() = modelDescription.modelName
-
-    /**
-     * Get the content of the modelDescription.xml file as a String
-     */
-    val modelDescriptionXml: String
-        get() = modelDescriptionFile.readText()
-
-    val modelDescription: ModelDescriptionProvider by lazy {
+    override val modelDescription: ModelDescriptionProvider by lazy {
         JaxbModelDescriptionParser.parse(modelDescriptionXml)
     }
 
@@ -113,12 +90,6 @@ class Fmu private constructor(
         return modelExchangeFmu
     }
 
-    /**
-     * Get the file handle for the modelDescription.xml file
-     */
-    private val modelDescriptionFile: File
-        get() = File(extractedFmu, MODEL_DESC)
-
     private val resourcesPath: String
         get() = "file:///${File(extractedFmu, RESOURCES_FOLDER)
                 .absolutePath.replace("\\", "/")}"
@@ -147,20 +118,8 @@ class Fmu private constructor(
                 fmiType, modelDescription.guid, resourcesPath, visible, loggingOn)
     }
 
-    override fun close() {
-        if (!isClosed.getAndSet(true)) {
 
-            LOG.debug("Closing FMU '$extractedFmu'..")
-
-            terminateInstances()
-            disposeNativeLibraries()
-            deleteExtractedFmuFolder()
-
-            fmus.remove(this)
-        }
-    }
-
-    private fun disposeNativeLibraries() {
+    override fun disposeNativeLibraries() {
         libraries.forEach {
             it.close()
         }
@@ -168,7 +127,7 @@ class Fmu private constructor(
         System.gc()
     }
 
-    private fun terminateInstances() {
+    override fun terminateInstances() {
         instances.forEach { instance ->
             if (!instance.isTerminated) {
                 instance.terminate()
@@ -180,60 +139,11 @@ class Fmu private constructor(
         instances.clear()
     }
 
-    /**
-     * Deletes the temporary folder where the FMU was extracted
-     */
-    private fun deleteExtractedFmuFolder(): Boolean {
-        if (extractedFmu.exists()) {
-            return extractedFmu.deleteRecursively().also { success ->
-                if (success) {
-                    LOG.debug("Deleted extracted FMU contents: $extractedFmu")
-                } else {
-                    LOG.debug("Failed to delete extracted FMU contents: $extractedFmu")
-                }
-            }
-        }
-        return true
-    }
-
-    override fun toString(): String {
-        return "Fmu(fmu=${extractedFmu.absolutePath})"
-    }
-
     companion object {
 
         private val LOG: Logger = LoggerFactory.getLogger(Fmu::class.java)
 
         private const val RESOURCES_FOLDER = "resources"
-        private const val BINARIES_FOLDER = "binaries"
-
-        private const val FMU_EXTENSION = "fmu"
-        private const val FMI4J_FILE_PREFIX = "fmi4j_"
-
-        private const val MODEL_DESC = "modelDescription.xml"
-
-        private val fmus = Collections.synchronizedList(mutableListOf<Fmu>())
-
-        init {
-            Runtime.getRuntime().addShutdownHook(Thread {
-                synchronized(fmus) {
-                    fmus.toMutableList().forEach {
-                        //mutableList because the list is modified during call
-                        it?.close()
-                    }
-                }
-            })
-        }
-
-        private fun createTempDir(fmuName: String): File {
-            return Files.createTempDirectory(FMI4J_FILE_PREFIX + fmuName).toFile().also {
-                File(it, RESOURCES_FOLDER).apply {
-                    if (!exists()) {
-                        mkdir()
-                    }
-                }
-            }
-        }
 
         /**
          * Creates an FMU from the provided File
