+++ fmi-import/src/test/kotlin/no/ntnu/ihb/fmi4j/importer/fmi1/Test.kt
@@ -13,9 +13,9 @@ class Test {
         Fmu.from(File(Test::class.java.classLoader.getResource("fmus/1.0/cs/BouncingBall.fmu").file)).use { fmu ->
 
             Assertions.assertEquals("{8c4e810f-3df3-4a00-8276-176fa3c9f003}", fmu.guid)
-            Assertions.assertEquals("BouncingBall", fmu.modelDescription.asCoSimulationModelDescription().attributes.modelIdentifier)
+            Assertions.assertEquals("BouncingBall", fmu.modelDescription.attributes.modelIdentifier)
 
-            fmu.asCoSimulationFmu().newInstance().use { slave ->
+            fmu.newInstance().use { slave ->
 
                 slave.setup()
                 Assertions.assertEquals(1.0, slave.readReal("h").value)
